Removes bench command
