+++ docs/installation.md
@@ -49,7 +49,6 @@ There are four ways to install Pilosa on MacOS: Use [Homebrew](https://brew.sh/)
       pilosa [command]
 
     Available Commands:
-      bench           Benchmark operations.
       check           Do a consistency check on a pilosa data file.
       config          Print the current configuration.
       export          Export data from pilosa.
@@ -108,7 +107,6 @@ There are four ways to install Pilosa on MacOS: Use [Homebrew](https://brew.sh/)
       pilosa [command]
 
     Available Commands:
-      bench           Benchmark operations.
       check           Do a consistency check on a pilosa data file.
       config          Print the current configuration.
       export          Export data from pilosa.
@@ -173,7 +171,6 @@ There are four ways to install Pilosa on MacOS: Use [Homebrew](https://brew.sh/)
       pilosa [command]
 
     Available Commands:
-      bench           Benchmark operations.
       check           Do a consistency check on a pilosa data file.
       config          Print the current configuration.
       export          Export data from pilosa.
@@ -262,7 +259,6 @@ There are three ways to install Pilosa on Linux: download the binary (recommende
       pilosa [command]
 
     Available Commands:
-      bench           Benchmark operations.
       check           Do a consistency check on a pilosa data file.
       config          Print the current configuration.
       export          Export data from pilosa.
@@ -327,7 +323,6 @@ There are three ways to install Pilosa on Linux: download the binary (recommende
       pilosa [command]
 
     Available Commands:
-      bench           Benchmark operations.
       check           Do a consistency check on a pilosa data file.
       config          Print the current configuration.
       export          Export data from pilosa.
